requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
sqlalchemy>=2.0
google-genai>=1.0
python-telegram-bot>=21.0
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0
browser-use>=0.12
playwright>=1.40
langchain-google-genai>=2.0
